rename to random component
