+++ CatLib.VS/CatLib/CatLib.csproj
@@ -212,14 +212,14 @@
     <Compile Include="Json\JsonProvider.cs" />
     <Compile Include="Json\JsonUtility.cs" />
     <Compile Include="Json\SimpleJsonAdapter.cs" />
-    <Compile Include="Maths\3rd\MathNet\Euclid.cs" />
-    <Compile Include="Maths\3rd\MathNet\Random\MersenneTwister.cs" />
-    <Compile Include="Maths\3rd\MathNet\Random\RandomSeed.cs" />
-    <Compile Include="Maths\3rd\MathNet\Random\RandomSource.cs" />
-    <Compile Include="Maths\3rd\MathNet\Random\Xorshift.cs" />
-    <Compile Include="Maths\MathProvider.cs" />
-    <Compile Include="Maths\RandomAdaptor.cs" />
-    <Compile Include="Maths\Math.cs" />
+    <Compile Include="Random\3rd\MathNet\Euclid.cs" />
+    <Compile Include="Random\3rd\MathNet\Random\MersenneTwister.cs" />
+    <Compile Include="Random\3rd\MathNet\Random\RandomSeed.cs" />
+    <Compile Include="Random\3rd\MathNet\Random\RandomSource.cs" />
+    <Compile Include="Random\3rd\MathNet\Random\Xorshift.cs" />
+    <Compile Include="Random\RandomProvider.cs" />
+    <Compile Include="Random\RandomAdaptor.cs" />
+    <Compile Include="Random\RandomFactory.cs" />
     <Compile Include="Routing\ExceptionRequest.cs" />
     <Compile Include="Routing\RouterEvents.cs" />
     <Compile Include="CatLib\Application.cs" />
