+++ CatLib.VS/CatLib/Random/3rd/MathNet/Euclid.cs
